fix: refactored the code
