+++ src/screens/manage_dao/_utils/types.tsx
@@ -1,0 +1,3 @@
+import { SettingsForm } from 'src/types'
+
+export type EditErrors = { [K in keyof SettingsForm]?: { error: string } };
